refactor(web): share PageIndex credential form
